Avoid erroring when computing 1000!^0.1
This was the original bug report that triggered all the discussion about
promotion to big floats. Unfortunately, I seem to be illiterate and
didn't realise that I'd misread it as 1000!*0.1 until rtoy pointed it
out.

More stupidly still, my changes for the former case actually end up
breaking this example, where exptrl assumed that (addk r1 0.0) was
always a float.

Maybe there's a better fix than this, avoiding any bigfloat conversion
when the result is small enough to fit in a machine float, but here's a
quick fix that at least gives a useful answer rather than a lisp error.