Re-add the plot_format "openmath" synonym for "xmaxima"
This synonym was silently dropped in Maxima v5.32.0, even though the
documentation says it should still work.  I imagine this was accidental
since a lot of the plotting functions were redone prior to this release.
See in particular commit 3066cbd7.

This removal broke (at least) the Embedded plotting in XMaxima (fixed
separately as commit 0fe27e05).