This is a fix for bug id #2696. After *prompt-prefix* and/or *prompt-suffix* are set, they are NOT inserted for askinteger() and asksign() on release 5.32.0, 5.32.1 and the latest git. They ARE, however, inserted for main prompt when set. This happens after the introduction of new format-prompt architecture in macsys.lisp.
Fix was necessary for both default-format-prompt function in
macsys.lisp and aformat function in mformt.lisp.

default-format-prompt function works now when stream is nil
or a real stream.

aformat function given nil as stream did not work as intended
by the original auther when control-string contained ~M directive.
Now this is fixed.