Bring Texinfo documentation into build system, by creating makefiles which execute makeinfo and build_index. Resulting .info and *-index.lisp files are included in this commit, although ideally they would be omitted since they are generated; however, src/sharefiles.mk is built before share, so they do not exist soon enough. The *-index.lisp files can be loaded and then their documentation items are findable via ? and ??. Ideally those index files would be loaded automatically.
In order to bring a new share subdirectory into this system, it is necessary to
create share/foo/Makefile.am and put foo on the SUBDIRS list in share/Makefile.am,
and add share/foo/Makefile to the list of makefiles to be generated in configure.in.
This is somewhat burdensome; I wonder if it can be simplified.