Extremely bare-bones documentation for clebsch_gordan contrib
This is just enough documentation to list the arguments to the functions
in the right order. It turns out that clebsh_gordan.tex seems to
describe an earlier version of the package, which unfortunately had a
slightly different API.

I'm leaving that TeX file for now, because it has references and
formulas that I haven't tried to transcribe into texinfo.