Be more paranoid about roots in DISCONTINUITIES-DENOM
Before this change, something of the form a^b was considered part of the
denominator if b was known to be negative. Now it is considered part of
the denominator unless b is known to be non-negative.

This results in more calls to SOLVE in REAL-ROOTS, which is
unfortunate, and then more calls to INTSUBS, because we evaluate the
definite integral on each piece. On the plus side, it fixes bug #2726.

I was a bit worried that this would slow things down massively, but the
test suite run times don't seem to change measurably here. (Phew!)