Fix for bug #2711 in geosum
Also, make it actually calculate the right thing. Before
geosum(0,-1,3) (which calculates sum(0*(-1)^n, n, 0, 3)) spat out an
error saying that the sum wasn't convergent.

Frankly, I don't really see the point of this sort of function: if the
user actually wants to sum something, he/she will use sum() and/or
simpsum. If the user is trying to learn first year analysis, he or she
should probably not blindly trust a CAS. Anyway, it's less broken than
it was. Yay.