Let bessel_j and bessel_i of integral order commute with conjugate
The Bessel functions J and I of integral order can commute with
conjugate even for arguments on the negative real axis.

The original code in commit e1f04ede (07/2006) let bessel_j commute
with conjugate when both the argument was off of the negative real
axis AND the order was an integer, which was very restictive.  Then
commit dda449be (02/2010) let bessel_j and bessel_i commute with
conjugate when the argument was off of the negative real axis,
regardless of the order; however, this is still too restictive.

Now bessel_j and bessel_i commute with conjugate when the argument is
off of the negative real axis OR the order is an integer.

The test suite runs fine with new tests.