Fix bug #2621: "gamma limit error". Thanks to Peter Bruin for the patch. From the bug report:
I finally discovered the reason for this bug: the series expansion
step in Gruntz's algorithm (see his thesis [1]) is not implemented
correctly. On page 49, it is stated that when expanding a logarithm,
log w must be replaced by h. In Gruntz's Maple implementation
(appendix of [1]), he does this by means of a magical declaration
ln(W) := e3; that is somehow picked up by Series().

In the Maxima implementation, the substitution log w -> h is only done
after expanding the whole expression. The problem is that by time,
the log w may have been transformed, e.g. because it appears inside an
exponential.

The patch limit-rewrite-logs.patch [see message below] fixes this,
at least for "pure exp-log functions" (the ones for which Gruntz's
algorithm was designed) by doing a suitable substitution on relevant
subexpressions of the form log(f(x)) before the series expansion.

The patch currently keeps the existing substition after the series
expansion because (1) if the expression involves other transcendental
functions whose Taylor expansion involves logarithms, we still want to
transform those if we see them, and (2) not doing this substitution
led to errors even in cases where it is apparently unnecessary,
e.g. in the test

integrate(x^6/(1 + x + x^2)^(15/2), x, 0, inf);

I hope this is an acceptable fix for this bug.

[1] D. Gruntz, On computing limits in a symbolic manipulation
system. Ph.D. thesis, ETH Zürich, 1996,
http://e-collection.library.ethz.ch/view/eth:40284