Uses gnuplot wxt terminal default options
Trying to change the properties of the wxt terminal causes errors in some
systems: those where there is no wxt installed and systems where a bug
prevents switching from qt to wxt. This commit restores the behaviour of
versions previous to 5.32, leaving the gnuplot default properties for
the wxt terminal; the only difference will be a different plot aspect ratio.
This fixes bugs #2683 and #2721.