Prevent some Lisp errors in featurep
When checking certain properties of an object with featurep, a Lisp
error would occur under several Lisps when the object was not a symbol.
This would happen for the 12 properties in OPERS (the default values in
opproperties), along with evfun, evflag, bindtest, nonarray and noun.

Now use SAFE-GET instead of GET when checking these properties.