Reorder APPROX-ALIKE to fix handling of CL arrays and hash tables
The apparent desired behavior of APPROX-ALIKE when given CL arrays or
hash tables is to descend and compare elements/entries.  The ordering
of the clauses in the COND form prevented this from working because
the test with ATOM came before the test with ARRAYP and HASH-TABLE-P.
Since CL arrays and hash tables are atoms, they were compared using
EQUAL:

(approx-alike #(1 2 3) #(1 2 3)) => nil
(approx-alike (make-hash-table) (make-hash-table)) => nil

Now the ARRAYP and HASH-TABLE-P clauses of the COND come before the
ATOM clause.  I've added a STRINGP clause, which compares strings with
STRING=, before the ARRAYP clause (so strings are treated the same as
before).

It doesn't appear that APPROX-ALIKE currently gets called with CL hash
tables or arrays (other than strings) during the test suite.  The test
suite runs fine (with no appreciable change in timing) under SBCL, GCL
and Clisp on Linux.

SBCL noticed some unreachable code in the ARRAYP clause.