Don't declare unused specialized DEFMETHOD parameters to be ignored
CLHS states that the macro expansion of DEFMETHOD "refers to" each
specialized parameter and that warnings do not occur if the body of
the method does not refer to specialized parameters.

GCL issued warnings calling the declarations redundant.  Apparently
old CMUCL and SBCL issued the same warning.  This warning comes from
Portable CommonLoops which these implementations have in common.