Remove an undeclared-but-used variable and a defined-but-unused variable
Commit ae4a8bf1 removed the declaration of the special variable ST from
suprv1.lisp, but a variable assignment was still performed on it.  The
rest of the code that used ST was already removed, so this assignment
was useless.  So remove the assignment.

Unrelated to the use of ST above, $BACKTRACE binds ST to *DEBUG-IO*
with LET before calling PRINT-ONE-FRAME in a loop.  Seemingly the
intention is to set the ST used in PRINT-ONE-FRAME.  Regardless of ST
being special or not, PRINT-ONE-FRAME specifies ST as an &aux variable
in its lambda list and binds it to *DEBUG-IO* anyway.  So remove the
binding in $BACKTRACE.