Initial cut at ~E and ~F formatters for bigfloat objects.
The intent is to be able to print bigfloat objects in f2cl'd code by
replace ~E with ~/bigfloat:format-e/ and ~F with ~/bigfloat:format-f/.
All parameters allowed by ~E and ~F are intended to be supported as
specified in the CLHS for floats.

More testing needed, but the basic functionality is supported for
bigfloat objects.

src/float.lisp:
o Implements the heart of ~E and ~F for maxima bfloat objects

src/numeric.lisp:
o Implement ~E and ~F via bigfloat::format-e and bigfloat::format-f,
  respectively, for bigfloat objects.