Make FPINTPART stricter
Don't try to compute the integer part of a big float if the quoted
precision isn't high enough that we actually know it. I realise that
this is a bit open to debate: after all, the floor of a (normal) float
is always defined. But bigfloats come with information about their
declared precision, so I reckon it makes sense to bail out if the answer
can't be calculated.

This commit was inspired by bug #2784. Running

  ceiling((207300647060*%e^-563501581931)/(403978495031*%e^-1098127402131));

ends up calling

  (FPENTIER
    '((BIGFLOAT SIMP 218)
      -359681197383905599631916620701909108192517867895676214965105532188
      771302019533))

Before this patch, we ended up trying to calculate

  (* -359681197383905599631916620701909108192517867895676214965105532188
     (expt 2 (- (- 218 771302019533))))

which is a ridiculously large number. In fact, SBCL on my machine dies
with a heap error because it tries to allocate 96412752432 bytes (89.8
GiB) to store the result(!).