Make PARTITION work when called on atoms
Before this patch, you got a lisp error if you accidentally passed
e.g. zero in to the function.

Also, the previous logic in $PARTITION for sums was really confusing (it
made a product and then passed it to PARTITION), and I can't believe it
made the code any faster (surely!). So this patch rewrites that code in
a way that I think is far easier to understand.