Match more strictly in INTEGRATE-EXP-SPECIAL
The function is basically an enormous COND and each clause was of the
same form, looking something like this:

  ((setq w (m2-exp-type-1a (facsum-exponent expr)))
   (let ((a (cdras 'a w))
         (c (cdras 'c w))
         (d (cdras 'd w))
         (r (cdras 'r w))
         (p (cdras 'p w))
         (v (cdras 'v w)))

     (do-stuff-here)))

I *think* the code assumed that if W was non-nil, then the resulting
alist would have a binding for all the variables we
expected. Unfortunately, this isn't quite true. For example, if expr is
one then m2-exp-type-4 binds D but not the other variables. This led to
an error when working on the example in bug #2781 (after fixing the
error in PARTITION with the previous patch).

Rather than put in all sorts of ugly lines asserting that the variables
are all non-nil, this patch moves the logic out to a separate macro
called SCHATCHEN-COND. Even with an extremely verbose description of the
macro, the result is still shorter than binding all the variables by
hand. (Let alone trying to do cleverer error checking!)