Avoid infinite recursion in SP1TRIGEX
There was an infinite recursion in SP1TRIGEX when evaluating

  powerseries(sin(x+qq),x,pp);

that caused bug #2772. It happened when trying to expand
sin(pp+qq) (which is free of X). Arguably, we might try to spot that the
expression is free of X before trying to expand it, but it's probably a
good idea to make sure that SP1TRIGEX can't ever trigger an infinite
recursion. The problem was that there was some special code to deal with
the case when all terms involved X, but no code for the opposite case.

This patch rewrites SP1TRIGEX in (I think) a more obvious fashion,
factoring out the actual expansion rules to avoid having to assign to
variables (and changing their meanings, which confused me a bit).

The rewrite uses SCHATCHEN-COND, a macro that I wrote recently to factor
some logic out of INTEGRATE-EXP-SPECIAL. As such, the patch moves that
macro from sin.lisp into schatc.lisp. There is no code change in the
macro, but there is a comment added about how to write an "otherwise"
clause.

The extra code responsible for actually fixing the bug is the ROTATEF
near the top of the new SP1TRIGEX.

I'm not including a test, because I don't know how to write one that
won't kill the testsuite on failure.