Modify SP1LOG2 to generate a definite integral
The previous code wrongly assumed that

  integrate(f'(x)/f(x), x) = log(f(x))

which is, of course, missing a constant of integration. Instead of
generating an indefinite integral to be expanded, generate a definite
integral.

It turns out that SP2INTEG2, which is responsible for expanding definite
integrals, is also broken, but that will be addressed in the next patch.