Correctly expand definite integrals in SP2INTEG2
The non-trivial code in SP2INTEG2 is for the case when we are trying to
expand

   integrate (f(u), u, a(x), b(x))

in x.

The previous code was just incorrect, as far as I can tell. The new
version expands a(x) and b(x) in x. If the results are not monomials in
x then we give up rather than trying to substitute arbitrary stuff into
a power series.

If they are monomials, we then expand the integrand (around zero: not
sure about that bit). If that works, integrating should be easy, since
it's just a power series. Next, substitute in the end points and
subtract the results.

Rather than adding another test, this fixes the test I added
yesterday. Oops...