Teach SP2EXPT to deal with sums in the exponent
In bug #2755, when you do

  powerseries (%e^x, x, 1);

you end up calling something like

  (SP2EXPT '((MEXPT SIMP) $%E ((MPLUS SIMP) 1 #:G242909)))

Because 1 + var is not a monomial in var, the previous version of the
code gave up. This patch teaches it to do the product if required.

Note that powerseries(a^x, x, 1) worked before because the exponent
appeared in the form x*log(a) which is a monomial in x.