Fix for bug #2760
This manifested itself as

    (%i3) powerseries(f(x),x,inf);
    (%o3) sum(at('diff(f(x)),1/x=0)/(i3!*x^i3),i3,0,inf)

but the problem was that SBSTPT unconditionally tried to substitute 1/x
back in for the expansion variable, even if the result looked like
diff(f(var), var). Of course, diff(f(1/x), 1/x) doesn't really make much
sense.

This patch just makes SBSTPT a bit more careful so it refuses to do this
with integrals and derivatives. If I've missed another class of function
where that wouldn't make sense, we can add it to INTDIFF-VARS-IN-EXPR
easily. (And possibly change the name!)