Make SRBINEXPND much more careful
The previous version blithely wrote a sum that is correct when
everything is nonzero, but wrong otherwise. Now we try to check if terms
are nonzero. If we have enough information, we can write various
versions of the binomial expansion. If the worst comes to the worst, we
give up, returning a noun form.