Tidy up call to SRBINEXPND for denominator
This is basically just a code tidy-up, but with a possible bug
fix. The LINPAT pattern in SRATEXPND has a constant coefficient on the
front. For the denominator case, the code explicitly divided by it after
expanding, basically doing

  expnd(1/(cc * f(x))) = 1/cc * expnd(1/f(x))

In the numerator case, I think we actually missed doing the
multiplication at all, but cc has always come out as 1 because we factor
out constant coefficients in advance.

Either way, it makes more sense to do the multiplication by cc in
SRBINEXPND ("expnd" above), which is what I assumed was happening when I
first read the code. This patch makes it work like that.