Remove some cruft from RISPLIT
realpart, imagpart and MABS have the REAL-VALUED property, so they do
not need to be a special case anymore.

erf has the COMMUTES-WITH-CONJUGATE property, so it does not need to be
a special case anymore.  The expressions produced by RISPLIT for erf
were often awkward anyway: the code was really old and so it expressed
this in terms of realpart and imagpart instead of conjugate.

The test suite runs fine.