Check some powerseries noun form results
These tests are just to make sure that we just return a noun form rather
than a wrong answer or explosion. I'm trying to get 100% branch coverage
across the code in series.lisp (other than the stuff to do with debug
prints), so want to exercise the "Give up and return a noun form"
branches.