Beef up psp2foldsum
Now it can deal with terms that are a mix of formal sums (as it expects)
and also non-sums. This means that it will successfully join up the
results of a partial fraction expansion that comes out like

   1/(x+1) + 1/(x+2) + 1

(or whatever). Before this patch, you'd get two formal sums: one for
each of the first two terms, because the "+1" confused things.

Also, factor out the "Take elements from a sequence and collect things
that match and things that don't" logic and put it in a function called
PARTITION-BY. This is now used by PARTITION (and thus $PARTITION) and
lives in csimp.lisp.