Removes @value{cedilha} from all texinfo documents
The older version of texinfo.tex previously used chocked with characters
@,{c} in the nodes. To work around that problem, a variable "cedilha" was
set, with a value of @,{c} that is replaced later on. That hack fixed the
problem for the creation of the info and html versions of the amnual, but
failed for the pdf version.

With the newer texinfo.tex, there is no need for that hack and if the nodes
use @,{c} instead of variable cedilha, the three versions of the manual are
successfully created.