Fixes a bug in describe and the HTML and PDF versions of the manual
There has been a long standing bug in the HTML and PDF versions of the manual,
where the index entries for a function or variable appeared several times
whenever 'deffnx' or 'defvrx' were used.

That problem did not affect describe in older versions, because describe
would only show the first ocurrance. In recent versions, described was
modified to show all occurrances and, for instance, in version 5.34.1
'? kill' would show a menu with 8 entries and all 8 entries would show the
same information. With this fix '? kill' will show directly the documentation
for 'kill', including the 8 different ways to use it. The result will also
look fine in the HTML and PDF versions of the manual.