Fix for bug #2854
This bug meant we got a hang from

    integrate(exp(-x)*sinh(sqrt(x)), x, 0, inf), domain: complex;

It happened in a limit computation in intsubs and was inside a lhospital
invocation. The regression (the computation hasn't always hanged) was
caused by commit 17ca33.

Before that commit, lhospital calculated a limit of the form (a+b)/c by
calling limit on ratsimp(a/c) + ratsimp(b/c). This limit can return a
function containing $UND, which caused a different bug, where a
L'Hospital's rule invocation would merrily return a result of the form
f(und)/g(und). Obviously, that wasn't great.

The "fix" for that bug was to instead call limit on something of the
form multthru(ratsimp((a+b)/c)). It seems that the bug went away because
the limit calculation that was returning $UND now failed and returned
NIL.

This patch re-instates the special behaviour for limits of quotient
whose numerator is a quotient (which fixes #2854). It also checks that
the result of the limit calculation is free of $UND terms, which seems
like a more principled way to fix the latter bug.

I'm not sure whether we should also fail on $IND terms: this patch
doesn't, but it would be pretty trivial to do so.