Avoid an error on limit(ind * inf)
This was triggered because we called $CSIGN from GETSIGNL inside a limit
calculation. Deep inside the sign calculation, it tries to calculate the
sign of $IND. If the special variable LIMITP is set, this just returns
$PNZ, otherwise it raises an error.

Unfortunately, $CSIGN was altered in commit c2bebf5 (in 2009) to bind
LIMITP to nil. To me, that seems like the wrong thing to do but I'm not
absolutely sure so I'm going for the "band-aid" approach and
binding *COMPLEX-SIGN* and calling $SIGN manually from the limit
calculation.

This no longer raises an error, but now SIMPINF-IC (my code, so my
fault!) returns a horrible result that looks like

   ind * limit(x, x, inf)

As a result, this patch also teaches SIMPINF-IC to return its original
argument if it hasn't improved as a result of calling $LIMIT on it.

I've also added a test to rtest_limit.mac which makes sure that
limit(ind*inf) doesn't raise an error.