Add a hook mechanism to PrMathML.lisp
This means that other code can register a hook to be called when the
MathML code is loaded (this is basically the same as Emacs's
eval-after-load, but without the syntactic sugar)

Using the hook, we can then get rid of the undefined function calls when
loading ezunits (which warn on at least SBCL). This also cuts the
apparent sort-of-dependency of ezunits on something in contrib.