Write a more helpful comment above SUBRESULT
I've been trying to figure out why the subresultant PRS algorithm
occasionally blows up (e.g. for the example Barton Willis identified at
the bottom of bug 125).

This is taking ages, because it really wasn't easy to work out the
relationship between the variables in Brown's paper and the program
variables. In particular h^1-d is h raised to the power of d minus
one (which, of course, is (1- d)). There went a good hour of my life...

This patch is basically to save the next poor soul from going through
the same tedious calculations.