Add a warning comment before SUBRESULT
This is an exhortation not to make the mistake that causes the second
half of bug 125 (evaluating "algsys([a+b+c=0, a*b*c=1], [a,b,c]),
algebraic : true" raises an error).

That mistake isn't fixed yet, because I don't know how, but at least
this comment might make the next example like it a bit easier to track
down.