Kill M-SUBSET in algsys.lisp and document SUBSETL
The M-SUBSET function was only used by SUBSETL in the same file. It
boiled down to calling the Common Lisp function SUBSETP with #'ALIKE1 as
the test. (Presumably this code predates the port to CL, since it
reinvents all of that)

This patch gets rid of M-SUBSET and also replaces the explicit iteration
in SUBSETL with a call to MEMBER-IF. On SBCL, at least, this gets open
coded to an iteration anyway.

Also get rid of an unnecessary (list nil), which conses up a new list
just to test it for equality (sadly, SBCL doesn't seem to be able to
optimise that away by itself).