Reformat and minimally document CALLAPPRS
This shortens an error message slightly, without removing any
information, so I think it's probably an improvement.

This patch also gets rid of a special variable called $ROOTSQUAD. It's
not used elsewhere in the Maxima source tree and a bit of archaeology
with git log -S suggests that it's been extraneous since the original
import!