Use NCONC instead of append in BAKALEVEL
The results of BAKALEVEL1 are all freshly consed, so we can be more
efficient here. Note: You wouldn't expect this to change, since
BAKALEVEL1 returns things that look like solutions to the equation ("x =
3" etc.), which obviously must be freshly consed because they weren't
known before the original call.

This patch also moves the documentation I wrote from BAKALEVEL1 to
BAKALEVEL and amends it to make sense.