Fix bug #2898: limit of continuous --> und
One way Maxima tries to find two-sided limits is to take the one-sided
limits from above and below and compare them.  In some of these cases
Maxima would find the correct one-sided limits but wouldn't recognize
that they are "the same" and so und would incorrectly be returned.

This changes BOTH-SIDE to compare the one-sided limits with MEQP rather
than ALIKE1.

The test suite runs fine, with new tests.