Try harder when translating literal exponents
This fixes bug 2934. The first two branches and the T branch in the
resulting cond are exactly the same as in the old code, but reformatted
a bit to make it easier to see what's going on. The third branch is new:
if we have an exponent A^B where B is a float, it checks to see whether
A is known to be non-negative. If so, it marks the result of the
calculation as being a float (because any non-negative number raised to
some floating number yields a float with the POWER function).

Note that we don't short-circuit when the base is known to be zero,
because we don't know that the calculation of the exponent is
side-effect free.

I've added a test to rtest_translator.mac, which doesn't get run as part
of the main testsuite. I assume that is because the tests are all a bit
chatty.