"Fix" RESTOREX in irinte.lisp
This function seems to have been designed to put a factor of x^k back
into an expression after it had been factored. The whole thing seems
pretty dubious: for a start, it didn't take an argument saying what "k"
to use. Then the function returned its argument unchanged if it was an
atom. I just can't see how this could possibly work.

Anyway, this patch alters RESTOREX so that it takes the power of the
variable that it needs to restore and so that it always actually
multiplies correctly by x^k.

I'm not really sure whether the resulting CAREFULFACTOR has any benefit
over just $FACTOR, but it no longer randomly divides its arguments by
stuff...

Testsuite passes fine with this patch, and we now return a noun form for
the integral given in bug 2933, rather than a wrong result.