Make $RESULTANT safer when $ALGEBRAIC is true
This patch was prompted by Barton Willis's follow-up comment in bug 125,
which pointed out that the equations caused algsys to explode if
algebraic was set to true.

After some debugging, I realised that the problem is caused by a call to
$resultant, which puts its arguments into CRE form. In this particular
case, the resulting kernels in VARLIST contain square roots. Suppose you
end up VARLIST containing x and "sqrt(x^10+1)" (in that order). Then,
when evaluating the resultant, you might square sqrt(x^10+1) (which has
degree 1 in the second variable), to get a term of degree 10 in the
first variable. This plays havoc with the resultant algorithm, which is
basically Euclid's algorithm and expects degrees to go down... The
result is a (confusing) lisp error.

Note that the user can quite easily get to a Lisp bug by calling
$resultant by hand, so we can't work around it in algsys. Instead, this
patch introduces some simple-minded code that tries to check the
variables in varlist are algebraically independent from one another. If
$algebraic is true and they are not known to be independent, the
$resultant code rebinds varlist to make sure we can't see something like
the example above happen.