Fix issues with gamma_incomplete's GRAD and psi's INTEGRAL properties
Part of the GRAD property for gamma_incomplete and the INTEGRAL
property for psi are implemented with Lisp lambda expressions.  The
parameters in the lambda lists should have been used to construct the
return values, but this was not done correctly (symbols matching the
parameters' names were used in quoted lists instead of using the
parameters' values).

The parameter names in the lambda lists match the dummy variable names
in the CARs of the GRAD and INTEGRAL properties, so the intended
functionality still happened to work.  If the dummy variables were
consistently renamed then this would cease to work.

Allegro CL noticed some unused variables