Remove the undocumented and broken grindef function
$GRINDEF would cause a Lisp error by trying to call the nonexistent
function GRINDEF.  Due to a call to EVAL, it could also cause a Lisp
error regarding unbound variables (depending on its arguments and the
Lisp implementation).

I don't know if there used to be a GRINDEF in Maxima or if $GRINDEF
was just calling a GRINDEF present in old Lisps.  If it used to be in
Maxima, then it looks like it was gone at least as far back as v5.0.

See also the (old) mailing list thread starting at
http://permalink.gmane.org/gmane.comp.mathematics.maxima.general/28167