Remove a redundant COND from MAPPLY1
The code was structured like

  (COND ((ATOM FN)
         (COND ((ATOM FN)
                ...)))
        ...)

so that the inner COND was unneeded.  Commit 66a62cf7 removed some old
clauses from this COND, but I'm not sure why there was the second ATOM
test anyway.

This basically removes a level of indentation from the code, while
also making the whitespace uniform and consistent.

No functional changes.  The test suite runs fine.