Avoid unnecessarily calling COMPLM in ORDER-LIMITS
COMPLM is used to compare the lower and upper limits of integration so
that we get the lower limit less than the upper limit.  If they are not
ALIKE1 then COMPLM calls limit and sign or asksign.  If we already know
that they are in the correct order, then don't bother with all of that.

Also get rid of an unneeded lexical variable.

The test suite runs fine.