Fix compile with recent GCL
I couldn't figure out when this happened from the GCL changelogs, but
the CLINES, DEFENTRY, OBJECT, QUIT and SYSTEM symbols have moved from
the LISP package to SYSTEM by at least 2.6.12.

In the other direction, *LOAD-PATHNAME* has been moved from SYS to
LISP (which now matches the standard).

Also, GET-SETF-METHOD is now an internal function in the SYSTEM
package (so system:get-setf-method doesn't work any more) but
GET-SETF-EXPANSION now exists (which means we can get rid of a "#+gcl")

This patch might break things with older GCLs, in which case we'll have
to do something cleverer, but I'm hoping we can get away with the stupid
approach because the symbols name macros so you can't just choose the
right one and funcall it...