Make get-setf-expansion work on GCL
Unfortunately, my previous patch broke older GCLs (e.g. 2.6.10) and
Robert's fix broke the GCL on my system (Debian unstable:
gcl_2.6.12-14). This version builds locally and I've also tested it with
the gcl_2.6.10-54 Debian package, so hopefully everything should build
correctly.

I vote we put as much GCL compatibility rubbish in the gcl-compat.lisp
file as possible, along with notes saying which versions it applies
to. That way, when we decide to drop support for ancient GCLs, we can
just go through that file deleting stuff.

Note that the testsuite is still broken with gcl_2.6.12-14 after this
patch :-(