If documentation can't be found, print a message.
If the byte-offset for the documentation isn't positive, print a
message.  This happens because build_index.pl wasn't able to find the
entry.  This usually means the texi file for node for the entry is
wrong in some way.

Now just print a message instead of lisp error about getting the
position from a directory.