Added code for crosscompiling VTK (with TCL support). (currently not enabled by default, use cmake -DUSE_VTK=YES .. to enable it)
Attention: The compilation time will increase!!!
The size of the installation will approximately be 100% more
than without VTK.)
Crosscompiliation & installation works now (and that was not
easy!), there is still an issue:
attempt to provide package vtk 6.3 failed: no version of
package vtk provided.
Perhaps somebody with more knowledge about tcl and vtk
knows, whats the problem?
I assume, that a path is wrong or a enviroment variable...