BIGFLOAT:EXPT must return a BIGFLOAT if one arg is.
Make sure (BIGFLOAT:EXPT a b) is a BIGFLOAT where a is a CL number and
b is a BIGFLOAT such that (= b (truncate b))

This shows up in bfloat(zeta(5)).  Without this fix, the answer is a
rational approximation to the true result.

Add a test for bfloat(zeta(5)) too.