Fix bug #3058: Lingering assumptions after an interrupted simpsum
Temporary assumptions were being leaked if sum (with simpsum true)
was interrupted (or otherwise aborted by transferring control
elsewhere).  The same thing would also happen with product (with
simpproduct true).

So wrap SIMPSUM1 and SIMPPROD1 in WITH-NEW-CONTEXT (which contains
an UNWIND-PROTECT to kill the context) to handle these assumptions.

The test suite runs fine.