Did forget these places when making plot+draw use individual files for every concurrent process.
Prior to commit f4af2a10e6471d1b9f9cd72c4aba887fa5c21621
if two instances of maxima were trying to plot at the same time this
would cause strange results since maxout.gnuplot, maxout.xmaxima and
maxout.dat would be accessed by both processes simultaneously.

Without the current commit a part of the change that fixed thiw was
missing. This would be causing no new trouble on systems that had
already used the affected functions with an older maxima version and
therefore was hard to find in ordinary tests. But if the functions
the current commit changes were called on a new system this could
under some circumstances lead to the files used for communication
with gnuplot not to be created - which would lead to an error
message saying that vtk is not installed even on systems where vtk
was up and running.

Thanks to Serge de Marre for doing the debugging on this part!