Try to exit gracefully if we are unable to connect to a server
Many Lisps invoke an error when unable to connect to the specified
port.  If *DEBUGGER-HOOK* is NIL then we would get dumped into the
debugger as expected.  The default debugger hook, after printing a
message, throws T with the tag RETURN-FROM-DEBUGGER.  However,
nothing would necessarily catch this and so we could get dumped
into the debugger in this case as well.

Instead, just ignore errors from OPEN-SOCKET in SETUP-CLIENT and
exit with a message if we cannot connect to the port.  This prevents
messy output like in the following example using Maxima's -s option
and SBCL:

$ ./maxima-local -s 31337
Connecting Maxima to server on port 31337
Maxima encountered a Lisp error:

 Socket error in "connect": ECONNREFUSED (Connection refused)

 Automatically continuing.
 To enable the Lisp debugger set *debugger-hook* to nil.

debugger invoked on a SB-INT:SIMPLE-CONTROL-ERROR in thread
#<THREAD "main thread" RUNNING {10077ECE23}>:
  attempt to THROW to a tag that does not exist: RETURN-FROM-DEBUGGER
[...debugger prompt...]