Remove control characters (repeated ^Z) from debugger output
I'm not sure what the repeated ^Z's in the output were intended to do.
I gave the debugger a try in XMaxima and the output caused an error
(in a dialogue box) about "not supported dirname" and I needed to
interrupt or restart Maxima to get things working again (I didn't
investigate further).  I also gave it a try in WxMaxima and it just
looked like garbage.