Return to Maxima from to_lisp REPL on EOF
I regularly use the EOF character to exit or otherwise control
various applications (including Maxima), but I noticed that the
to_lisp REPL did not attempt to check for an EOF and so a Lisp
error would occur if one was seen.

Now we're sent back to the Maxima REPL on EOF.  Another option
would be to exit completely (for example by calling $QUIT instead
of TO-MAXIMA), but I think going back to Maxima makes more sense
and is the safer option.