Revert commit b88ba7
This commit removed the repeated control character ^Z from the
debugger output because I noticed breakage (described below).
However, these characters are useful when running the debugger
in the GNU Emacs dbl shell.

On my current system (a "fairly standard" Linux box) I see the
following when these control characters are printed:

* In xmaxima the characters cause an error (in a dialogue box) about
  "not supported dirname"
* In wxmaxima the characters look like "garbage"
* In imaxima the characters are printed as "^Z"
* Different terminals print the characters differently (several
  that I tried do not print anything noticable, while one that I
  tried prints them in a funny way).

So now the issues above appear again but the dbl shell behavior is
back.  Something needs to be done to fix these issues, but I'm not
yet sure what is best.