Fix bug #481: ('m)[1] (meval)
In MAPPLY1 if the operator is MQUOTE then the ARRAY flag (as given by
the special variable ARYP) was being ignored when constructing the new
expression from the quoted expression.

This caused

(%i1) ('m)[1];
(%o1) m(1)

instead of the desired behavior

(%i3) ('m)[1];
(%o3) m[1]

The test suite runs fine, with new tests.