Add --enable-mathjax configure option.
Add configure option --enable-mathjax to enable usage of MathJax in
the html docs,  This just loads the MathJax script in each html file.

To support this, a new texinfo flag, mathjax, is added.  When MathJax
is enabled, the flag is set so that the texi file can use

@ifset mathjax
@html
MathJax formula
@end html
@end ifset

to enable MathJax formulas.  If mathjax is not set, the formula is not
inserted into the html.  Perhaps best usage is

@ifset mathjax
@html
MathJax formula
@end html
@end ifset

@ifclear mathjax
@math{...}
@end ifclear