Improve the message printed for setchecked variables
This changes
  <var> SET TO <val>
to
  <var> is being set to <val>

I think the latter is more correct since the message is printed
before the assignment actually takes place (this is noticeable
at the break prompt provided when setcheckbreak is true).

This also further changes the format of the message so that
superfluous backslashes are not printed when display2d is false:

(%i2) display2d: false$
display2d SET TO false
(%i3) x: 'foo$

x\ SET\ TO\ foo

The new message is consistent with the message printed when
refcheck is true (see also commit 1a8b4c).